Reduce buffer size from multiples of 16 bits to 8 bits.
Not a problem, but overkill.

+++ common/shader_utils.js
@@ -130,7 +130,7 @@ function bind_indices(gl , indices , buffer)
     gl.bufferData
     (
 	gl.ELEMENT_ARRAY_BUFFER ,
-	new Uint16Array(indices) ,
+	new Uint8Array(indices) ,
 	gl.STATIC_DRAW
     );
     return buffer;
